pyproject.toml
[project]
name = "pubscout"
version = "0.1.0"
description = "CLI-driven publication scanning agent — fetches, filters, scores, and reports new research papers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "arxiv>=2.1.3",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "jinja2>=3.1",
    "rapidfuzz>=3.0",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
pubscout = "pubscout.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pubscout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
